+++ src/main/java/be/baur/sds/content/AbstractStringType.java
@@ -1,17 +1,18 @@
 package be.baur.sds.content;
 
 import be.baur.sds.MixedType;
+import be.baur.sds.SDS;
 import be.baur.sds.common.NaturalInterval;
 
 /**
  * An {@code AbstractStringType} represents an SDA node with string content and
- * a minimum/maximum length. Unlike other types, it is null-able by default.
+ * a minimum and maximum length. Unlike other types, it is null-able by default.
  * <br>
  * See also {@link StringType} and {@link BinaryType}.
  */
 public abstract class AbstractStringType extends MixedType {
 
-	private NaturalInterval length = null; // null means any length is allowed.
+	private NaturalInterval length = NaturalInterval.INFINITE; // default is to allow any length
 
 
 	/**
@@ -26,10 +27,9 @@ public abstract class AbstractStringType extends MixedType {
 
 
 	/**
-	 * Returns the allowed length interval. This method returns null if any length
-	 * is allowed.
+	 * Returns the allowed length interval. This method never returns null.
 	 * 
-	 * @return a natural interval, may be null
+	 * @return a natural interval, not null
 	 */
 	public NaturalInterval getLength() {
 		return length;
@@ -37,32 +37,12 @@ public abstract class AbstractStringType extends MixedType {
 
 
 	/**
-	 * Sets the allowed length interval. A length of null means any length is
-	 * allowed.
+	 * Sets the allowed length interval. This method does not accept null.
 	 * 
-	 * @param length a natural interval, may be null
+	 * @param length a natural interval, not null
+	 * @throws IllegalArgumentException is length is null
 	 */
 	public void setLength(NaturalInterval length) {
-		this.length = length;
-	}
-
-
-	/**
-	 * Returns the minimum content length.
-	 * 
-	 * @return a minimum length
-	 */
-	public int minLength() { // do we really need this method?
-		return length != null ? length.min : 0;
-	}
-
-
-	/**
-	 * Returns the maximum content length.
-	 * 
-	 * @return a maximum length
-	 */
-	public int maxLength() { // do we really need this method?
-		return length != null ? length.max : Integer.MAX_VALUE;
+		this.length = SDS.requireNonNull(length, "length must not be null");
 	}
 }